WIP: trying to fix wireguard connection
Details: - everything inside the VPC can access the docker apps inside
the app instance but wireguard is unable to route the traffic properly.
This was a working module from an old project so I know it should have
worked out of the box, but I guess something is wrong somewhere. Still
trying to fix it.